requirements.txt
pymongo
python-dotenv
langchain
langchain-core
langchain-community
faiss-cpu
sentence-transformers
torch
tqdm
huggingface-hub
transformers
datasets

